requirements.txt
Django==5.2
djangorestframework
djangorestframework-simplejwt
djongo
pandas
requests
textblob
